+++ src/icemac/ab/importer/importer.py
@@ -1,10 +1,8 @@
-# -*- coding: utf-8 -*-
 import zope.container.btree
 import zope.interface
 import icemac.ab.importer.interfaces
 
 
+@zope.interface.implementer(icemac.ab.importer.interfaces.IImporter)
 class Importer(zope.container.btree.BTreeContainer):
     """Importer containing files for import."""
-
-    zope.interface.implements(icemac.ab.importer.interfaces.IImporter)